refactor rect border packages and co-locate shadow modifiers
