+++ app/src/main/java/com/skul/yuriy/composeplayground/feature/animatedBorderRect/border/multilayer/MultiLayerRectShadowBox.kt
@@ -1,4 +1,4 @@
-package com.skul.yuriy.composeplayground.feature.animatedBorderRect.border
+package com.skul.yuriy.composeplayground.feature.animatedBorderRect.border.multilayer
 
 import androidx.compose.animation.core.animateDpAsState
 import androidx.compose.animation.core.tween
@@ -16,47 +16,36 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.input.pointer.pointerInput
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
-import com.skul.yuriy.composeplayground.util.shadowborder.rect.drawOutlineBlurMaskShadow
 
 @Composable
-fun BlurredRectShadowBox(
+fun MultiLayerRectShadowBox(
     modifier: Modifier = Modifier,
     color: Color,
     cornerRadius: Dp,
-    initialBlurRadius: Dp,
-    pressedBlurRadius: Dp,
-    initialHaloShadowWidth: Dp,
-    pressedHaloShadowWidth: Dp
+    initialHaloBorderWidth: Dp,
+    pressedHaloBorderWidth: Dp
 ) {
     var isPressed by remember { mutableStateOf(false) }
 
-    val animatedBlurRadius by animateDpAsState(
-        targetValue = if (isPressed) pressedBlurRadius else initialBlurRadius,
-        animationSpec = tween(durationMillis = 300),
-        label = ""
-    )
-
-    val animatedHaloBorderSize by animateDpAsState(
-        targetValue = if (isPressed) pressedHaloShadowWidth else initialHaloShadowWidth,
+    val animatedSpread by animateDpAsState(
+        targetValue = if (isPressed) pressedHaloBorderWidth else initialHaloBorderWidth,
         animationSpec = tween(durationMillis = 300),
         label = ""
     )
 
     Box(
         modifier = modifier
-            .drawOutlineBlurMaskShadow(
-                color = color.copy(alpha = 0.7f),
-                haloBorderWidth = animatedHaloBorderSize,
-                cornerRadius = cornerRadius,
-                blurRadius = animatedBlurRadius
+            .drawOutlineMultiLayerShadow(
+                color = color.copy(alpha = 0.6f),
+                haloBorderWidth = animatedSpread,
+                cornerRadius = cornerRadius
             )
             .then(
                 if (isPressed) {
-                    Modifier.drawOutlineBlurMaskShadow(
+                    Modifier.drawOutlineMultiLayerShadow(
                         color = color,
                         haloBorderWidth = 4.dp,
-                        cornerRadius = cornerRadius,
-                        blurRadius = 2.dp
+                        cornerRadius = cornerRadius
                     )
                 } else {
                     Modifier
