+++ app/src/main/java/com/skul/yuriy/composeplayground/feature/animatedBorderRect/border/shadowlayer/ShadowLayerRectShadowBox.kt
@@ -1,4 +1,4 @@
-package com.skul.yuriy.composeplayground.feature.animatedBorderRect.border
+package com.skul.yuriy.composeplayground.feature.animatedBorderRect.border.shadowlayer
 
 import androidx.compose.animation.core.animateDpAsState
 import androidx.compose.animation.core.tween
@@ -16,7 +16,6 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.input.pointer.pointerInput
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
-import com.skul.yuriy.composeplayground.util.shadowborder.rect.drawOutlineShapeShadowByShadowLayer
 
 @Composable
 fun ShadowLayerRectShadowBox(
@@ -36,14 +35,14 @@ fun ShadowLayerRectShadowBox(
 
     Box(
         modifier = modifier
-            .drawOutlineShapeShadowByShadowLayer(
+            .drawOutlineShadowLayerShadow(
                 color = color,
                 haloBorderWidth = animatedSpread,
                 cornerRadius = cornerRadius
             )
             .then(
                 if (isPressed) {
-                    Modifier.drawOutlineShapeShadowByShadowLayer(
+                    Modifier.drawOutlineShadowLayerShadow(
                         color = color,
                         haloBorderWidth = 4.dp,
                         cornerRadius = cornerRadius
